Refactor helper methods
This renames and refactors the following methods: flattenTree and listToTree
The name flattenTree was inaccurate as the method is used to build the tree
array without flattening, hence renames it as objToTree. While the name
listToTree is renamed as treeToObject since it is used as a reverse method
of objToTree.
Regarding the refactorization, the initial solution of listToTree is to
convert changed parameters to object. However, this solution wasn't robust
and it causes issue when parameter object has depth level of nested object
more than 1. e.g. log_throttling.count.something.
Though, log_throttling.count works fine but the solution was complex and
hard to read.

The new solution uses ancestor's originalValue of changed node to update
new values. This ensures log_throttling.count.something's value is updated.