Move main routing functions into MXS_FILTER SESSION
Moving the routeQuery and clientReply functions into MXS_FILTER_SESSION
and using the same session type for both routers and filters allows the
routing code to directly call the routeQuery method of the upstream or
downstream object.

This change simplifies things but it breaks any code that does not
implement the APIs properly, that is, filter sessions that aren't a
mxs::FilterSession and router sessions that aren't a mxs::RouterSession.

Now that the object can be directly called, there's no need for the
various types that contain the instances and sessions and combine them
into a callable object.