Return mxs::FilterSession from newSession
By returning the filter session class from the newSession function, we
remove the need for the freeSession function. This also removes the need
for the upstream and downstream parameters in the newSession function as
the filter session creation code can directly call the setUpstream and
setDownstream functions of the returned object.

This commit breaks all modules that do not derive from the
mxs::FilterSession class. The following filters are affected and must
eventually be converted to use the filter template:

* luafilter
* regexfilter
* topfilter
* tpmfilter