Add clearable props for select-dropdown
clearable of selected items should be controlled by props.
It was using multiple props value to control clearable ability
of v-select. If multiple selection is false, it adds clearable button. However,
there is case that single selection shouldn't be clearable. For example:
service relationship when creating a listener, empty item is not allowed.

This adds clearable props to select-dropdown and to components which
are using select-dropdown.