MXS-3107 At each monitor loop, check cluster configuration
When dynamic node detection is enabled we check at each monitor
loop whether the cluster configuration has changed and act
accordingly.

There are two cases; all nodes agree what the configuration is
(when a node is removed, its configuration will be updated) or
they are of different opinion. Now only the former case is
handled.