Add constant color palettes to vuetify theme
Constant colors variables should be added to vuetify theme configuration.
This relieves the usage of custom color classes.
This also cleans up css style in data-table, move its styles to the
component instead of having it in main.scss. Replaces all hard-coded
colors with colors variables or color classes. This helps to add
dark-mode in the future easier.