Fix core parameter alteration
The core parameters weren't validated before the configuration was
attempted which hit a debug assertion. This was prompted by the REST API
unit tests that PATCH the modified object back instead of putting back
only modified values.

Converted the unknown parameter warning into an error as unknown
parameters stop the alteration process.

The directory parameters are stripped from the JSON before being given to
the core specification and configuration. This is a temporary workaround
until the directory parameters are added into the specification.