Prevent double-free in cache
The fact that the weak_ptr can be locked implies that the session is still
running and the `this` pointer captured in the lambda is valid. This is
not a neat solution and is done only to prevent the double free. A
reworking of the synchronization should be done at some point.