Refactor the way of controlling dialog visibility
base-dialog was using computed props to control visibility.
This requires passing onClose and onCancel handlers from parent
to control two-way data bindings. This approach isn't necessary
and cause redundant and repetitive handlers. In addition, if
persistent props is not used in v-dialog and user press ESC key
to close dialog, it breaks two-way data bindings.

This refactors base-dialog to emit 'input' event to control dialog
visibility instead of using computed props and handler function props.
For components such as select-dialog, confirm-dialog and forms, it's
better to use ref to control visiblity of base-dialog as those components
are used as a wrapper of base-dialog.
This refactorization will breaks some components and pages unit tests.
The next commit will fix those tests.