Configure only updated values
This prevents the on_set handler for being called for no-op changes. This
can occur when a resource is first retrieved with a GET request and then
subsequently PATCHed back with only a subset of the parameters updated.

Converting the value to JSON is a generic method of detecting equality but
it is not the most efficient one. Comparing the JSON value to the native
type would be more effective but until this turns out to be a problem, the
JSON conversion is adequate.