MXS-3178: Implement WebSocket connections
The WebSocket class handles the concrete task of framing the messages
according to the WebSocket protocol and doing event handling.

The callback provided to the create function is the provisional method of
delivering data. It will be changed later when the need to close the
WebSocket early arises.

The WebSocket::shutdown must be called when the REST API is being shut
down to make sure all the connections are closed before the libmicrohttpd
daemon is stopped.