MXS-3154 Add loopFetchOlderLogs
This function fetches older logs than current until the log content div
is scrollable. This allows user to scroll up to get old logs if
current logs received are to small which make it unable to scroll.
This may happen when log_source is maxlog and log_debug=1 as
multiple log lines in maxscale is now ignored.