MXS-3128: Make listener configurable at runtime
Storing the listener parameters inside a mxs::WorkerGlobal makes it easier
to handle the modification of the shared_ptr stored in the listener. By
reconstructing the shared data from the configuration object, the process
of reconfiguration is greatly simplified.

This also makes the configuration alteration behave in a
semi-transactional manner; a failure to configure the Configuration won't
take the new parameters into use.