MXS-3235: Fix insertstream
The commit after an insert would not be detected as a LOAD DATA LOCAL
INFILE would be ongoing when it is processed by the client protocol. This
is a more complex problem that stems from the linkage of the state
tracking for LOAD DATA LOCAL INFILE for client and backend protocols.