MXS-3204 Load MariaDBProtocol automatically
Several modules already compile-time linked with mariadbclient,
making runtime loading redundant. The main executable now also links
with the protocol library, so that all modules can assume that the
library is always available.

Also changed the name of the module to be more descriptive.

The MXS_MODULE object is still available in the module handler so that
it is listed in e.g. "maxctrl list modules".