MXS-1687: Track cluster ownership
The cluster ownership that is decided by the monitor implementation is now
exposed to the rest of the system. This is an improvement of the current
passive mechanism in MaxScale that allows the active status of a Maxscale
instance to be determined at the cluster level.

This mechanism doesn't prevent multiple actions from taking place and is
not something that can be used to build a system with exactly-once
semantics. It does, however, provide a way to make the avrorouter and
kafkacdc routers highly available.