MXS-3208 Prepare for queued requests
When a response from the backend server is received we must
be able to send not only that response but also responses of
queued up requests whose responses may be generated immediately.
Thus, the Mongo instance needs the dcb.

Also, as the downstream component will not change and as it will
be needed when executing queued requests it is passed to the Mongo
instance at construction time.