MXS-1687: Only replicate from owned clusters
The Replicator class is used by both avrorouter and kafkacdc which make
both highly available when cooperative monitoring is used.

The fact that the check is done periodically leaves a window during which
multiple MaxScale instances can replicate at the same time. Given that the
purpose of this feature is to provide HA for MaxScale itself, this is
adequate.

Implementing exactly-once processing of replication depends on the module
that handles the event processing. As the avrorouter provides a service,
it is not needed there. Kafkacdc would benefit greatly from it but since
the problem involves communication between multiple components, it is not
easy.