MXS-3208 Remove problematic member variable
Before the change, making a mxsmongo::Packet from any of the
derived classes would leave the m_pData pointer pointing at the
wrong place; if a derived class instance would have been created
from the copy data would be read from the wrong place. As the
member is not needed, it is better to remove it.