MXS-3208 Introduce mxsmongo::Database::Command
Each Command class is intended to handle a particular Mongo
command. The idea is that for any particular Mongo command a
class will be created that then is capable of performing that
command. So, adding support for new commands is merely a question
of creating that class and adding it to a lookup table, but nothing
else need to be changed.