MXS-3249 Fix mxs2729_rebalance system test
The correspondence between number of connections and load is not
as direct as was expected. Now, instead of expecting that all
connections will be distributed evenly across all threads, we
simply compare the load of the threads and are happy once it
reaches the balance we told MaxScale to reach. The maximum
number of rounds is still based upon MaxScale having to distribute
all connections evenly between the threads.