MXS-1923: Add extra state information to servers
The server diagnostic now contains the information provided the
monitor. This helps convey the extra information to the user but keeps it
separate from the information that's relevant to routing.

With this change, the Slave of External Server, Master Stickiness and
Binlog Relay states can be moved to the monitor that assigns them.