MXS-3292: Don't update the initial database
The database the client used to connect to MaxScale should always be used
whenever backend connections are created. This keeps the starting
conditions for all backend connections the same regardless of the point at
which they were created.

There was also a race condition where the change of the database could
happen before the backend server handshake was read causing the initial
database to change to something other than the intended one.