qc_sqlite: Fix asserts
The original intention was that first the setup function must
be called and only then the process and thread init functions.
However, there's no need for that and, in fact, it's more
"natural" to perform process/thread initialiation first.
Further, the assert() had no effect a so it seems NDEBUG
is defined also in Debug builds.