MXS-3108: Add filters to session
These filters are applied before any filters that are defined for the
topmost service. This is a relatively straightforward way of handling it
from the API viewpoint as sessions already had a placeholder for filter
relations. Injecting filters into the downstream components of the routing
chain is more complex as they are not exposed via the API.

The filter sessions are created whenever the filter relationships are
changed. This means that altering the filter chain of a session causes the
filter sessions to be reopened. This can be optimized to only open new
sessions for filters that were added. In addition, the code currently does
no checking on the session state which means that a filter session could
start in the middle of a query.