MXS-3305: Add log level configuration to sessions
This allows extra log levels to be enabled for sessions at runtime. The
main purpose of this is to enable verbose logging for long-running
sessions. As the session ID is the same as the connection ID sent to the
client, this allows test programs to enable log_info for a subset of the
connections it creates. This should greatly help reduce the verbosity of
logs in tests where extra information is needed but there is a lot of
traffic.

The same mechanism could be used to conditionally enable logging based on
some other configuration parameter. For example, the username and host
combination of a client could be used to only log messages from a known
"bad" host.