MXS-3300 Support and validate nested parameters
Nested parameters are parameters that are not evaluated and used
by the top-level object, but by another object that indirectly
is loaded by the first. In practice the other object is a module
(e.g. a protocol module) on which the top-level object (e.g. a
listener) depends.

In config files, nested parameters are expressed using dot-notation,
where the scope must be the value of a top-level parameter referring
to a module. For instance, with a listener section like

  [MyListener]
  type=listener
  protocol=mongodb
  ...

where the value of 'protocol' is a module, parameters for 'mongodb'
are expressed like

  mongodb.user=hello
  mongodb.password=world

During validation, it will be verified that the scoped parameters
indeed refer to a module. If that's the case, then the module will
be loaded and its specification be used for validating the actual
parameters, which will be provided without that scope. That is,
like

  user=hello
  password=world

During configuration, the nested parameters will be provided to the
post_configure()-method as separate mxs::ConfigParameters objects
indexed by the module name they are intended for.

If the parameters are provided using a JSON object, then the
nested parameters are expressed as fields in an object whose
key is the name of the module in question. For instance,

  {
    ...
    protocol: "mongodb",
    mongodb: {
      user: "hello",
      password: "world"
    }
  }

During configuration, the nested JSON object will be converted into
a mxs::ConfigParameters to provide to the post_configure()-method.