MXS-3108: Add is_idle method to ProtocolConnection
This is needed to make sure that the session being modified is completely
idle. This guarantees that all added filter sessions will start from a
known state where no unexpected results are delivered.

This new method can also be used to make the `ping` method simpler by
calling it first before calling the `ping` method of the backend
connection.