MXS-3313 Use timeout in redis and memcached storages
A timeout is now used when connecting to Redis and Memcached.

In the case of Redis the effect is that in case Redis is not available,
the connecting will not simply hang with the eventual effect of MaxScale
crashing when it attempt so close the connection. From a client's point
of view the session creation will simply fail.

In the case of Memcached, that connects only when an operation is actually
performed, the session creation will succeed, but then every operation
will take 'timeout' time to execute as we will first wait for memcached
before going to the backend.