MXS-1245: Move routing information into RouteInfo
This allows the RouteInfo class to represent the state of the query
classifier which allows peeking the next state without actually modifying
the active query classification state. This also allows the routing code
in readwritesplit to be organized so that the target finding code is
completely separate from the actual routing part.