MXS-1245: Reorganize RouteInfo initialization
Moving the default value initialization into the variable declaration
removes the need for custom constructors. Added explicit default copy and
move constructors.

Moved documentation from the member variables in to the getter
methods. This way the documentation is in the public member functions
instead of the private member variables.

Grouped setters and getters in groups to make it easier for users to see
what information is available. As the setters are only used by the
QueryClassifier, they could be made private and QueryClassifier made a
friend class.