MXS-3316 Connect to Redis in a secondary thread
Connecting to Redis is now done in a secondary thread. The effect is
that neither the routing thread, nor the session will be blocked by
the operation. The routing thread will not have to wait for the
operation to finish and the session will simply not use the cache
until the operation has finished.