MXS-3316 Transparently (re)connect to Redis
If the storage looses the connection to Redis, caching will simply
be disabled and a reconnection attempt will be made at regular
intervals. So, if Redis is available, caching will be used, and if
it is not, caching will not be used, and the situation may change
during the lifetime of a session.