MXS-3327 Prepare for pinging memcached
So that a client will not become unusable in case it is no possible
to reach the memcached server, we must be aware of whether it is
so that we in that case can bybass the cache.

This commit prepares for making those modifiations; the memcached
token is now aware of whether it is connected or not, whether a
(re)connection attempt is in process and when a connection was
established.