MXS-1245: Add streaming of queries to the master
The master now accepts queries even without waiting for the first result
to arrive. This is done in an optimistic fashion as the expected use-case
for this is bulk loads of data. If these statements are done inside of a
transaction, they can still be replayed but the replay will execute them
one at a time.

The type of the query that is to be streamed can also be a SELECT
statement that returns a resultset. This can happen if, for example, a
batch of INSERT ... RETURNING statements is executed and the generated
results are returned. Sadly, due to the fact that the C connector only
supports streaming of SQL that doesn't generate resultsets, this will be
hard to test.

As the new code is simply a performance improvement, it's hard to
functionally test that the queries are actually streamed and the new code
is taken into use. The remaining options for verification are to either
measure timing or parse the logs.