MXS-3327 Gracefully handle unreachable memcached
The reachablity of Memcached is now checked and in an
asynchronous fashion. Consequently, neither the routing thread
nor the client will be stalled while that is taking place. The
client will simply not used cached data until the presence of
the memcached server has been checked.

If an error occurs during the communication with memecached,
the cache is taken out of use and the client will access the
backend directly until the memcached server is reachable again.