MXS-1245: Separate master selection from replacement
The potential new master must be returned as the routing target for
TARGET_MASTER type queries. This prevents more queries from being sent to
the old master and allows doing of the replacement only when we actually
start routing queries to it.

Moving the statistics counter incrementation out from the target selection
fixes the case where the statistics would be counted twice for queries
that are first queued and then routed. This is also much more clear as the
target selection code does only one thing: selects a target.