MXS-3334: Set trx target when routing the first query
This prevents the case where the START TRANSACTION being sent to the
master fails and we don't detect that a replay should start. This also
makes the whole transaction target tracking clearer as we know where we're
attempting to route it instead of only knowing where we ended up routing
it.