MXS-3338: Fix ending of read-only transactions
The read-only transactions that failed during the execution of the commit
would not trigger the transaction replay as m_target_node was cleared
before the result for the COMMIT was read.

This fix also seems to have unearthed a bug where read-only transactions
would hang if a result was expected from the backend. A separate test case
for this needs to be added.