MXS-3339: Fix execution of queued session commands
The code assumed that all session commands generated a response which
breaks when a COM_STMT_CLOSE is put into the history.

In theory this assumption could still be held if the COM_STMT_PREPARE
command matching the COM_STMT_CLOSE is erased from the history. As doing
this would bring savings in memory usage as well as speed up the
reconnection process, it might be worth doing eventually.

Improved info level logging of session commands to also display the
command being executed. This should help solve problems that arise from
session command execution of non-text commands.