Document grants for xpandmon and csmon
Made the documentation of xpandmon consistent with the csmon documentation
and removed the list style declaration of tables that need grants. Other
grants in MaxScale are listed as a set of GRANT commands and the xpandmon
should follow this style (also easier to copy-paste).

Changed usernames and passwords to the same in all examples. This should
help if/when they are changed.