MXS-3342: Persistent connections cause signal 11 when no matching connection is found
There is no guarantee that the persistent pool contains a connection which matches our host
when proxy protocol is un use. We should handle this.

I am contributing the new code of the whole pull request, including one or
several files that are either new files or modified ones, under the BSD-new
license.