Make special command parsing optional
If a module in the routing chain requires query classification, the client
protocol code can reuse the query classification result to determine
whether the custom parsers need to be used. This speeds up the execution
of all normal commands.

Exposed the query classification requirement as a routing capability bit
and added it to the relevant routers that already use the query
classifier.