Make the ignored query mechanism simpler
Now that the COM_CHANGE_USER processing is moved out from the normal
routing code, the query result ignoring mechanism can be implemented by
simply freeing the result instead of routing it.

In the current codebase the query result ignoring is only used by the
backend MariaDB protocol to generate COM_PING packets that are ignored. A
separate state for this would allow the whole query result ignoring
mechanism to be removed.