MXS-3371 Handle bsoncxx exceptions
bsoncxx throws exceptions, for instance, if you attempt to convert
something to a type that it is not. As that may happen only if
the client sends a broken command it simplifies the code if we simply
catch those exceptions instead of all over the place checking that
something really is of the type expected (and specified).