MXS-3371 Add support for Insert
This commit adds support for the MongoDB command insert.

This commit also introduces a slight change in focus. Instead of
providing MongoDB access to existing relational tables, the
aim is simply to provide a NoSQL interface to MariaDB. In practice
this means that MongoDB documents will be stored to tables having
just two columns; 'id' and 'doc' where the id of the document is
stored to the former and the document as JSON string to the latter.
Query support will depend upon what is possible to provide using
JSON-functions of MariaDB, possibly extended with processing inside
MaxScale.