Fix PS direct execution in readwritesplit
A binary protocol command whose ID we didn't know would be classified as
something that should be routed to the master. This classification is
correct but the information about the direct execution was lost in the
process. This information is needed to detect that the binary protocol
command should not be routed by readwritesplit until the prepared
statement has completed. With the addition of the streaming of queries to
the master, the direct execution would end up receiving the error
readwritesplit sends for all unknown PS IDs.