MXS-3375: Remove history deduplication
The history compression would break the execution of prepared statements
as the ID stored in the buffer wouldn't be correct. Since the benefits of
the history deduplication are quite limited, removing it should have no
measurable effect in practice.

Improved the code that handles unknown prepared statement IDs. Calling
clientReply() inside the write() function of the backend protocol can
cause logic errors in routers that change their state only after calling
the routeQuery() method of their endpoints. Also removed the delivery of
errors for COM_STMT_CLOSE as it doesn't generate a response.