MXS-3290 Allow client protocol module to specify capabilities
In case a non-MariaDB protocol module is using the MariaDB backend
protocol, it may have requirements that the MariaDB client protocol
does not have. Now it is possible for a client protocol to specify
capabilities to be taken in consideration by the backend protocol
when it routes packets from the server towards the client.