MXS-3362: Track the number of expected responses
Since the client protocol expects a specific response after a state
change, it must wait until all but the last response is sent. This happens
when a normal query and one of the special commands is sent to MaxScale at
the same time.

For example, if the following is executed with mysql_send_query(), it
would cause the client protocol to think that the USE command completed
successfully:

  SELECT SLEEP(1);
  USE database_that_does_not_exist;

It appears that if some commands require special treatment, then all
commands and their responses must be tracked.