MXS-3362: Store response types in mxs::Backend
With it, the knowledge about what was routed to
the backend and what is expected from it is explicitly known which makes
things simpler to comprehend.

The IGNORE_RESPONSE response_type allows routers to send both session
commands and normal commands to the backends without waiting for the
responses individually. This way the clientReply routing code is reduced
to checking whether the result should be routed to the client or
discarded.

The separation of responses also greatly simplifies error handling by only
requiring one type of check without a separate check for whether the
backend was executing a session command that should be ignored or whether
it was a query and the client is expecting a result.