MXS-3362: Replace session commands in readwritesplit
Replaced the storage and queuing of session commands in readwritesplit
with immediate execution of session commands. This improves both the
performance of the router and the simplicity of the code by using the
information stored inside the mxs::Backend.

Removed the separate storage of session commands that were done inside
transactions. Doing it was a pessimistic approach as the only case where
it would prevent extra commands from being executed on the slaves was when
a transaction failed due to the master going down. Executing the session
commands when they arrive reduces the latency between the slaves executing
the commands and them being available for read operations.