MXS-3265 Separate DCB::close() into different functions depending on use case
The concrete type of a dcb is known whenever close() is called, so
the different types can have dedicated close-functions. The backend
dcb case is also separated into the send-to-pool and the normal close
cases.

Using different functions increases clarity, as the base class
function no longer needs to handle every case. The cleanup is not
yet final.