MXS-3362: Check responses to session commands
The backend protocol module now checks whether the response that it just
sent upstream is the same as the accepted response that is sent to the
client and if it's not it closes the connection.

For responses that arrive too early, the check is postponed until the next
query or response arrives. This isn't optimal but, due to the problems
with pooled connections, this is the only reliable way to implement it
now.

One limitation of using the history as the source of validation is that
when history pruning is implemented, it is possible that a backend is so
slow that it doesn't get to execute the session command before it is
removed from the history. If this happens, the code currently closes the
connection to be safe. Another solution would be to leave the connection
open and assume the server is OK. Since a slow server usually is not
something that's wanted, closing the connection should be OK.