MXS-3362: Fix protocol state changes
It is possible that multiple queries are in progress while a transition
from ROUTING to PREPARING_PS is made in the backend. The corresponding
transition back to ROUTING must only be done when every expected response
has been received. Otherwise the execution of the statements goes out of
sync and queries that refer to prepared statements will fail as the ID
isn't found.

In addition, the delayed packets were written without checking whether a
state change was made. This was a known possibility but it turns out to be
a bigger problem than was expected.

For the client protocol, a similar problem existed where the command was
put into the history before it fully completed. This only happened with a
COM_STMT_PREPARE with numerous parameters.