MXS-3362: Ignore erased commands
As the COM_STMT_PREPARE commands aren't stored in the history, any binary
protocol commands that refer to an erased prepared statement must skip the
result comparison. This is not a very good solution but it is an
acceptable compromise while other parts of the system are changed and
testing is done.