MXS-3362: Fix multi-result response comparison
A multi-statement query with a `USE <database>` as the first command would
be stored in the history. This caused the query validation to be done
twice for the query and if there was more than one type of result, it
would cause the validation to fail.

Also removed the redundant connecting to services that don't exist from
the test case.