MXS-3362: Fix writing of large COM_STMT_PREPARE
When executing a COM_STMT_PREPARE that spans multiple packets, the backend
wouldn't route the latter part of it when it should. In addition, the
command byte was always extracted regardless of whether the current packet
container it or not. Doing the large packet tracking for each packet
separately from the actual query result tracking solves it.