MXS-3362: Add support for large session commands
This removes the limitation where a session command that spans multiple
packets would prevent the history from being used.

With the changes done to the session command system, we must know when
we're expecting more data for a recorded command. Once we know that,
adding support for large session commands is a one line fix that appends
the data to m_pending_cmd.

Separate states for LARGE_HISTORY_PACKET and LARGE_PACKET seem to be more
in line with the idea of a finite state machine that only looks at the
current state to know what to do next. The LARGE_PACKET state could do
what LARGE_HISTORY_PACKET does but it would have to inspect the previous
state to be able to handle commands that are being recorded.

Also adjusted the LARGE_PACKET state to bypass any checks that were done
by route_statement() and route the query directly. This prevents any
processing from taking place for the tail end of the packet.