MXS-3362: Move sescmd configuration into service
Moved the session command configuration into the service and remove it
from readwritesplit. This causes other routers that don't support it to
accept the parameters but this will be handled later on in a separate
commit.

The service parameters still use the old style configuration which should
be changed over to the new system. In addition, this commit changes the
default value of prune_sescmd_history to true. This was already documented
in the upgrading document but it still has to be documented in the general
service parameter section.

Currently, disabling the session command history won't prevent it from
being collected in memory but this is actually beneficial as the result
comparison uses it. If the history would not be collected, a separate
mechanism for storing the result information would have to be
created. Eventually this might be what has to be done as having a very
small history can cause problems with the new implementation.