MXS-3362: Retain all sescmd responses in memory
Using the history as the source of accepted answers won't work that well
as it complicates the checking of the results and places unwanted
limitations. This does create a problem with an ever increasing history.

The pruning of the response history can't be done until the backend
protocol can reliably communicate with the client protocol via the
protocol data object. Currently, the backend protocol cannot interact with
the protocol data inside its destructor as the session pointer, and by
extension the protocol data, is invalid if it's in the pool.

This commit fixes the mxs1787_slave_reconnection test.