MXS-3362: Make sescmd history optional
The routers now need to declare a capability in order for the session
command history to be recorded. This prevents it from being done for
routers that don't require it, namely readconnroute.

Moved the incrementation of the next ID to be done inside
record_for_history. This doesn't affect the functionality as a routing
failure will cause the session to be closed.