MXS-3362: Use callbacks to compare sescmd results
Using callbacks removes the need to deliver hangup events to backends that
delivered a wrong response before the accepted response was received. This
clarifies the behavior by causing the connections to be directly closed
with the correct error message. In addition, it removes the unnecessary
checking of responses whenever data is routed to the backends.

With the previous implementation, the error would be delivered only after
the hangup event triggered a history replay. It would then detected the
conflict and report the error. This was a less efficient solution as it
required two handleError calls and a full reconnection cycle in order for
the mismatch to be reported.