Make filter construction uniform
The runtime creation now uses the raw JSON values which allows more
complex parameter types to be used in filters.

In addition, the parameter validation is now done on both the module as
well as the core level. This also means that the module must be stored in
the parameters but, due to the REST API layout, the module isn't naturally
in the parameters and thus must be injected into it.