Add built-in default timeout for tests
This should help detect cases where tests hang in places where they don't
expect to hang. Currently if this happens, the CTest timeout would kill
the process and the logs would not be copied.

The default value for the timeout is half of the CTest timeout. This might
need to be adjusted for any tests that are expected to take longer than
this.