MXS-3401: Use mxs::config::Configuration in SERVICE
As a side-effect of this change, the service now stores the servers it
uses in two forms: as a list of strings as well as a list of actual
servers, the former of which isn't used after it has been configured. This
is only temporary and can be replaced with actual pointers to the objects
once a parameter type for it is added.

Configuration changes done at runtime also won't change it as they operate
directly on the servers themselves. This currently cause problems with if
the service changes from one type of targets to another e.g. from servers
to a cluster.