Add symbol location checks to module loading
Added back the check that makes sure the module being loaded is actually
the one that defines the module entry point symbol. This prevents the case
where the GUI would get errors about non-module libraries being loaded
when a new object was being created.

In addition, it guarantees that the module declaration function is only
called once. This isn't something that should be guaranteed but currently
it must be as modules do actual work in the declaration functions.